// src/models/bbsinfo.ts
export interface IBbSinfo{
    nrc:string
    cursoid: string 
    pago: string 
    tipo: string 
    calificable: string 
    patron: string 
    periodo: string 
    id_alumno: string 
    id_inst: string 
    zonal: string 
    instregistrado:    string
    userregistrado: string
    usuariovisiblecurso:string
    supregistrado: string

}